pyproject.toml
[project]
name = "wordepth"
version = "0.1.0"
description = "WorDepth: Monocular depth estimation with language guidance"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "absl-py==2.0.0",
    "autopep8==2.0.4",
    "cachetools==5.3.2",
    "certifi==2023.11.17",
    "charset-normalizer==3.3.2",
    "cmake==3.27.7",
    "contourpy==1.1.1",
    "cycler==0.12.1",
    "filelock==3.13.1",
    "fonttools==4.49.0",
    "fsspec==2023.10.0",
    "ftfy==6.1.3",
    "google-auth==2.23.4",
    "google-auth-oauthlib==1.0.0",
    "grpcio==1.59.3",
    "huggingface-hub==0.19.4",
    "idna==3.4",
    "importlib-metadata==6.8.0",
    "importlib-resources==6.1.3",
    "Jinja2==3.1.2",
    "kiwisolver==1.4.5",
    "lit==17.0.5",
    "Markdown==3.5.1",
    "MarkupSafe==2.1.3",
    "matplotlib==3.7.5",
    "mpmath==1.3.0",
    "multidict==6.0.4",
    "networkx==3.1",
    "numpy==1.24.4",
    "nvidia-cublas-cu11==11.10.3.66",
    "nvidia-cublas-cu12==12.1.3.1",
    "nvidia-cuda-cupti-cu11==11.7.101",
    "nvidia-cuda-cupti-cu12==12.1.105",
    "nvidia-cuda-nvrtc-cu11==11.7.99",
    "nvidia-cuda-nvrtc-cu12==12.1.105",
    "nvidia-cuda-runtime-cu11==11.7.99",
    "nvidia-cuda-runtime-cu12==12.1.105",
    "nvidia-cudnn-cu11==8.5.0.96",
    "nvidia-cudnn-cu12==8.9.2.26",
    "nvidia-cufft-cu11==10.9.0.58",
    "nvidia-cufft-cu12==11.0.2.54",
    "nvidia-curand-cu11==10.2.10.91",
    "nvidia-curand-cu12==10.3.2.106",
    "nvidia-cusolver-cu11==11.4.0.1",
    "nvidia-cusolver-cu12==11.4.5.107",
    "nvidia-cusparse-cu11==11.7.4.91",
    "nvidia-cusparse-cu12==12.1.0.106",
    "nvidia-nccl-cu11==2.14.3",
    "nvidia-nccl-cu12==2.18.1",
    "nvidia-nvjitlink-cu12==12.3.101",
    "nvidia-nvtx-cu11==11.7.91",
    "nvidia-nvtx-cu12==12.1.105",
    "oauthlib==3.2.2",
    "opencv-python==4.9.0.80",
    "packaging==23.2",
    "Pillow==10.1.0",
    "platformdirs==4.0.0",
    "protobuf==4.25.1",
    "pyasn1==0.5.1",
    "pyasn1-modules==0.3.0",
    "pycodestyle==2.11.1",
    "pyparsing==3.1.2",
    "python-dateutil==2.9.0.post0",
    "PyYAML==6.0.1",
    "regex==2023.10.3",
    "requests==2.31.0",
    "requests-oauthlib==1.3.1",
    "rsa==4.9",
    "safetensors==0.4.0",
    "scipy==1.10.1",
    "six==1.16.0",
    "sympy==1.12",
    "tensorboard==2.14.0",
    "tensorboard-data-server==0.7.2",
    "tensorboardX==2.6.2.2",
    "timm==0.9.10",
    "tokenizers==0.13.3",
    "tomli==2.0.1",
    "torch==2.0.1",
    "torchvision==0.15.2",
    "tqdm==4.66.1",
    "traitlets==5.11.2",
    "transformers==4.27.4",
    "triton==2.0.0",
    "typing-extensions==4.7.1",
    "tzdata==2023.3",
    "urllib3==2.0.4",
    "wcwidth==0.2.13",
    "werkzeug==3.0.0",
    "xxhash==3.4.1",
    "yapf==0.40.1",
    "yarl==1.9.2",
    "zipp==3.16.2",
    "setuptools>=65.0.0,<69",
    "ultralytics>=8.4.15",
]
